fiexed delivery updates

+++ src/services/package.ts
@@ -1,14 +1,23 @@
 import PackageModel, { IPackage } from "../models/package";
 import DeliveryModel, { IDelivery } from "../models/delivery";
 
-// Get all packages
 const getAllPackages = async (): Promise<IPackage[]> => {
   return await PackageModel.find();
 };
 
-// Get package by ID
-const getPackageById = async (packageId: string): Promise<IPackage | null> => {
-  return await PackageModel.findOne({ packageId });
+const getPackageById = async (
+  packageId: string
+): Promise<{ package: IPackage | null; delivery?: any }> => {
+  const pkg = await PackageModel.findOne({ packageId });
+
+  if (pkg && pkg.activeDeliveryId) {
+    const delivery = await DeliveryModel.findOne({
+      deliveryId: pkg.activeDeliveryId,
+    });
+    return { package: pkg, delivery };
+  }
+
+  return { package: pkg };
 };
 
 // Create a new package
@@ -17,30 +26,48 @@ const createPackage = async (packageData: IPackage): Promise<IPackage> => {
   return await newPackage.save();
 };
 // Create many new packages
-const createManyPackages = async (packagesData: IPackage[]): Promise<IPackage[]> => {
-  const packageIds = packagesData.map(pkg => pkg.packageId);
-  const existingPackages = await PackageModel.find({ packageId: { $in: packageIds } });
+const createManyPackages = async (
+  packagesData: IPackage[]
+): Promise<IPackage[]> => {
+  const packageIds = packagesData.map((pkg) => pkg.packageId);
+  const existingPackages = await PackageModel.find({
+    packageId: { $in: packageIds },
+  });
 
   if (existingPackages.length > 0) {
-    throw new Error('Some packageIds already exist: ' + existingPackages.map(pkg => pkg.packageId).join(', '));
+    throw new Error(
+      "Some packageIds already exist: " +
+        existingPackages.map((pkg) => pkg.packageId).join(", ")
+    );
   }
   const createdPackages = await PackageModel.insertMany(packagesData);
   return createdPackages;
 };
 
-// Update a package by ID
 const updatePackage = async (
   packageId: string,
   updateData: Partial<IPackage>
 ): Promise<IPackage | null> => {
-  console.log({updateData})
-  return await PackageModel.findOneAndUpdate({ packageId }, updateData, {
-    new: true,
-  });
+
+  const updatedData = await PackageModel.findOneAndUpdate(
+    { packageId },
+    { $set: updateData },
+    { new: true, upsert: true }
+  );
+  return updatedData;
 };
 
-// Delete a package by ID
 const deletePackage = async (packageId: string): Promise<IPackage | null> => {
+  const pkg = await PackageModel.findOne({ packageId });
+
+  if (!pkg) {
+    throw new Error("Package not found");
+  }
+  if (pkg.activeDeliveryId) {
+    // Delete the associated delivery
+    await DeliveryModel.findOneAndDelete({ deliveryId: pkg.activeDeliveryId });
+  }
+
   return await PackageModel.findOneAndDelete({ packageId });
 };
